Sorted React state update on unmounted component
