+++ components/organism/MyLinks.jsx
@@ -1,4 +1,4 @@
-import Pagination from "../Pagination";
+import Pagination from "./Pagination";
 
 const MyLinks = () => {
   return (
